requirements.txt
opencv-python
djitellopy
requests
boto3
pyqt5
python-dotenv